Installer will now not proceed if ignition not provided, useful for debugging, Added extra labels to image, removed installer env containerfile

+++ assets/install-rhcos.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+IGNITION="/tmp/bf.ign"
+
 default_device=/dev/mmcblk0
 if [ -b /dev/nvme0n1 ]; then
   default_device="/dev/$(cd /sys/block; /bin/ls -1d nvme* | sort -n | tail -1)"
@@ -22,31 +24,30 @@ if [ -e "${boot_fifo_path}" ]; then
   offset=$(strings -a -t d ${cfg_file} | grep -m 1 "7zXZ" | awk '{print $1}')
   if [ -s "${cfg_file}" -a ."${offset}" != ."1" ]; then
     echo "INFO: Found text" | tee /dev/kmsg
-    cat ${cfg_file} | tr -d '\0' > /tmp/bf.ign
+    cat ${cfg_file} | tr -d '\0' > $IGNITION
     ls -lah $cfg_file | tee /dev/kmsg
   fi
   rm -f $cfg_file
 fi
 
-if [ ! -f /tmp/bf.ign ]; then
-  echo "INFO: Creating ignition file" | tee /dev/kmsg
-  echo '{"ignition": {"version": "3.4.0"}}' > /tmp/bf.ign
+if [[ -f "$IGNITION" ]] && jq -e . "$IGNITION" >/dev/null 2>&1; then
+  echo "INFO: Valid ignition file found – proceeding with installation." | tee /dev/kmsg
+
+  coreos-installer install "$device" \
+    --append-karg "console=hvc0 console=ttyAMA0 earlycon=pl011,0x13010000 ignore_loglevel" \
+    --ignition-file "$IGNITION" \
+    --offline
+
+  sync
+
+  if efibootmgr -v | grep -q "Red-Hat CoreOS GRUB"; then
+    efibootmgr -b $(efibootmgr -v | grep "Red-Hat CoreOS GRUB" | awk '{print $1}' | cut -d' ' -f1) -B
+  fi
+
+  efibootmgr -c -d $device -p 2 -l '\EFI\redhat\grubaa64.efi' -L "Red-Hat CoreOS GRUB"
+
+  reboot
+else
+  echo "INFO: Ignition file missing or invalid, skipping installation." | tee /dev/kmsg
+  exit 1
 fi
-
-
-dd if=/dev/zero of=$device bs=1M count=1 status=progress
-
-coreos-installer install $device \
-  --append-karg "console=hvc0 console=ttyAMA0 earlycon=pl011,0x13010000 ignore_loglevel" \
-  --ignition-file /tmp/bf.ign \
-  --offline
-
-sync
-
-if efibootmgr -v | grep -q "Red-Hat CoreOS GRUB"; then
-  efibootmgr -b $(efibootmgr -v | grep "Red-Hat CoreOS GRUB" | awk '{print $1}' | cut -d' ' -f1) -B
-fi
-
-efibootmgr -c -d $device -p 2 -l '\EFI\redhat\grubaa64.efi' -L "Red-Hat CoreOS GRUB"
-
-reboot
